4.30.17

+++ bot.js
@@ -27,20 +27,22 @@ function hasRole(mem, role) {
 }
 
 client.on('ready', () => {
-    console.log('Tylbot EXPERIMENTAL is ONLINE');
+    console.log('TylBot EXPERIMENTAL is ONLINE')
 });
 
-client.on('ready', () =>{
-  client.user.setGame('very unstably.')
-});
+// client.on('ready', () =>{
+//   client.user.setGame('very unstably.')
+// });
 
 client.on('message', message => {
     var deletewords = ['>say ', '>roast ', '>8ball ','>arrest '];
     var i = 0
     while (i < deletewords.length) {
-        if (message.content.toLowerCase().includes(deletewords[i])) {
+      if(hasRole(message.member, "Member")); {
+        if(message.content.toLowerCase().includes(deletewords[i])) {
             message.delete();
         }
+      }
         i++;
     }
 });
@@ -48,9 +50,9 @@ client.on('message', message => {
 client.on('message', message => {
   var args = message.content.split(/[ ]+/);
   if(commandIs("say", message)){
-      if (hasRole(message.member, "Tyrant Hoang") || hasRole(message.member, "admin")) {
+      if (hasRole(message.member, "Tyrant Hoang")) {
           if (args.length === 1) {
-              message.channel.sendMessage('No argument was defined. Usage: ``>say [message to say]``');
+              message.channel.sendMessage('No argument was defined. Usage: ``>say [message to say]``');
           } else {
               message.channel.sendMessage(args.join(" ").substring(5));
           }
@@ -60,8 +62,8 @@ client.on('message', message => {
   }
   if(commandIs("purge", message)){
       if(hasRole(message.member, "Moderator") || hasRole(message.member, "Tyrant Hoang") || hasRole(message.member, "admin")){
-          if(args.length >= 3){
-              message.channel.sendMessage('Too many arguments were defined. Usage: ``>purge [number of messages to delete]');
+          if(args.length >= 3) {
+              message.channel.sendMessage('Too many arguments were defined. Usage: ``>purge [number of messages to delete]');
           } else {
               var msg;
               if(args.length === 1){
